src/main/java/models/Department.java
package models;

import java.util.Objects;
import java.util.ArrayList;


public class Department {
    private String name;
    private int id;
    private ArrayList<Section> sections;


    public Department(){}

    public Department(String name) {
        this.name = name;
    }

    public String getName() { return name; }
    public void setName(String name) { this.name = name; }
    public int getId() { return id; }
    public void setId(int id) { this.id = id; }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Department department = (Department) o;
        return id == department.id &&
                Objects.equals(name, department.name);
    }

    @Override
    public int hashCode() {
        return Objects.hash(name, id);
    }
}